MAplot {affyPLM}	R Documentation

Relative M vs. A plots

Description

Create boxplots of M or M vs A plots. Where M is determined relative to a specified chip or to a pseudo-median reference chip.

Arguments

... 	 Additional parameters for the routine

A 	 A vector to plot along the horizonal axis

M 	 A vector to plot along vertical axis

subset 	 A set of indices to use when drawing the loess curve

show.statistics 	 If true some summary statistics of the M values are drawn

span 	 span to be used for loess fit.

family.loess 	 "guassian" or "symmetric" as in loess.

cex 	 Size of text when writing summary statistics on plot

See Also

mva.pairs

[Package affyPLM version 1.16.0 Index]