PLMset-class {affyPLM}	R Documentation

Class PLMset

Description

This is a class representation for Probe level Linear Models fitted to Affymetrix GeneChip probe level data.

Objects from the Class

Objects can be created using the function fitPLM

Slots

probe.coefs:

Object of class "matrix". Contains model coefficients related to probe effects.

se.probe.coefs:

Object of class "matrix". Contains standard error estimates for the probe coefficients.

chip.coefs:

Object of class "matrix". Contains model coefficients related to chip (or chip level) effects for each fit.

se.chip.coefs:

Object of class "matrix". Contains standard error estimates for the chip coefficients.

model.description:

Object of class "character". This string describes the probe level model fitted.

weights:

List of objects of class "matrix". Contains probe weights for each fit. The matrix has columns for chips and rows are probes.

phenoData:

Object of class "phenoData" This is an instance of class phenoData containing the patient (or case) level data. The columns of the pData slot of this entity represent variables and the rows represent patients or cases.

annotation

A character string identifying the annotation that may be used for the ExpressionSet instance.

description:

Object of class "MIAME". For compatibility with previous version of this class description can also be a "character". The class characterOrMIAME has been defined just for this.

cdfName:

A character string giving the name of the cdfFile.

nrow:

Object of class "numeric". Number of rows in chip.

ncol:

Object of class "numeric". Number of cols in chip.

notes:

Object of class "character" Vector of explanatory text.

varcov:

Object of class "list". A list of variance/covariance matrices.

residualSE:

Object of class "matrix". Contains residual standard error and df.

residuals:

List of objects of class "matrix". Contains residuals from model fit (if stored).

Methods

weights<-

signature(object = "PLMset"): replaces the weights.

weights

signature(object = "PLMset"): extracts the model fit weights.

coefs<-

signature(object = "PLMset"): replaces the chip coefs.

coefs

signature(object = "PLMset"): extracts the chip coefs.

se

signature(object = "PLMset"): extracts the standard error estimates of the chip coefs.

se<-

signature(object = "PLMset"): replaces the standard error estimates of the chip coefs.

coefs.probe

signature(object = "PLMset"): extracts the probe coefs.

se.probe

signature(object = "PLMset"): extracts the standard error estimates of the probe coefs.

coefs.const

signature(object = "PLMset"): extracts the intercept coefs.

se.const

signature(object = "PLMset"): extracts the standard error estimates of the intercept coefs.

getCdfInfo

signature(object = "PLMset"): retrieve the environment that defines the location of probes by probe set.

image

signature(x = "PLMset"): creates an image of the robust linear model fit weights for each sample.

indexProbes

signature(object = "PLMset", which = "character"): returns a list with locations of the probes in each probe set. The list names defines the probe set names. which can be "pm", "mm", or "both". If "both" then perfect match locations are given followed by mismatch locations.

Mbox

signature(object = "PLMset"): gives a boxplot of M's for each chip. The M's are computed relative to a "median" chip.

normvec

signature(x = "PLMset"): will return the normalization vector (if it has been stored).

residSE

signature(x = "PLMset"): will return the residual SE (if it has been stored).

boxplot

signature(x = "PLMset"): Boxplot of Normalized Unscaled Standard Errors (NUSE).

NUSE

signature(x = "PLMset") : Boxplot of Normalized Unscaled Standard Errors (NUSE) or NUSE values.

RLE|

signature(x = "PLMset") : Relative Log Expression boxplot or values.

Note

This class is better described in the vignette.

Author(s)

B. M. Bolstad bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley.

[Package affyPLM version 1.16.0 Index]