PLMset2exprSet {affyPLM}	R Documentation

Convert a PLMset to an ExpressionSet

Description

This function converts a PLMset to an ExpressionSet. This is often useful since many Bioconductor functions operate on ExpressionSet objects.

Usage


PLMset2exprSet(pset)
pset2eset(pset)

Arguments

pset 	 The PLMset to convert to ExpressionSet.

Details

These functions convert PLMset objects to ExpressionSet objects. This is often useful since many Bioconductor functions operate on ExpressionSet objects. Note that the function pset2eset is a wrapper for PLMset2exprSet.

Value

returns a ExpressionSet

Author(s)

Ben Bolstad bmb@bmbolstad.com

See Also

ExpressionSet

Examples


data(affybatch.example)
Pset <- fitPLM(affybatch.example)
eset <- pset2eset(Pset)

[Package affyPLM version 1.16.0 Index]