bg.correct.LESN {affyPLM}	R Documentation

LESN - Low End Signal is Noise Background corrections

Description

This function background corrects PM probe data using LESN - Low End Signal is Noise concepts.

Usage


bg.correct.LESN(object,method = 2,baseline = 0.25, theta=4)

Arguments

object 	 an AffyBatch

method 	 an integer code specifying which method to use

baseline 	 A baseline value to use

theta 	 A parameter used in the background correction process

Details

This method will be more formally documented at a later date.

The basic concept is to consider that the lowest end of intensites is most likely just noise (and should be heavily corrected) and the highest end signals are most likely signal and should have little adjustment. Low end signals are made much smaller while high end signals get less adjustment relative adjustment.

Value

An AffyBatch

Author(s)

Ben Bolstad bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley.

Examples


data(affybatch.example)
affybatch.example.bgcorrect <- bg.correct.LESN(affybatch.example)

[Package affyPLM version 1.16.0 Index]