normalize.ExpressionSet {affyPLM}	R Documentation

Normalization applied to ExpressionSets

Description

Allows the user to apply normalization routines to ExpressionSets.

Usage


  normalize.ExpressionSet.quantiles(eset,transfn=c("none","log","antilog"))
  normalize.ExpressionSet.loess(eset,transfn=c("none","log","antilog"),...)
  normalize.ExpressionSet.contrasts(eset, span = 2/3, choose.subset = TRUE, subset.size = 5000, verbose = TRUE, family = "symmetric",transfn=c("none","log","antilog"))
  normalize.ExpressionSet.qspline(eset,transfn=c("none","log","antilog"),...)
  normalize.ExpressionSet.invariantset(eset,prd.td = c(0.003, 0.007),verbose=FALSE,transfn=c("none","log","antilog"),baseline.type=c("mean","median","pseudo-mean","pseudo-median"))
normalize.ExpressionSet.scaling(eset,trim=0.02,baseline=-1,transfn=c("none","log","antilog"))

Arguments

eset 	 An ExpressionSet

span 	 parameter to be passed to the function loess.

choose.subset

subset.size

verbose 	 verbosity flag

family 	 parameter to be passed to the function loess.

prd.td 	 cutoff parameter (details in the bibliographic reference)

trim 	 How much to trim from the top and bottom before computing the mean when using the scaling normalization

baseline 	 Index of array to use as baseline, negative values (-1,-2,-3,-4) control different baseline selection methods

transfn 	 Transform the ExpressionSet before normalizing. Useful when dealing with expression values that are log-scale

baseline.type 	 A method of selecting the baseline array

... 	 Additional parameters that may be passed to the normalization routine

Details

This function carries out normalization of expression values. In general you should either normalize at the probe level or at the expression value level, not both.

Typing normalize.ExpressionSet.methods should give you a list of methods that you may use. note that you can also use the normalize function on ExpressionSets. Use method to select the normalization method.

Value

A normalized ExpressionSet.

Author(s)

Ben Bolstad, bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley.

See Also

normalize

Examples


data(affybatch.example)
eset <- rma(affybatch.example,normalize=FALSE,background=FALSE)
normalize(eset)

[Package affyPLM version 1.16.0 Index]