normalize.scaling {affyPLM}	R Documentation

Scaling normalization

Description

Allows the user to apply scaling normalization.

Usage


normalize.scaling(X,trim=0.02,baseline=-1,log.scalefactors=FALSE)
normalize.AffyBatch.scaling(abatch,type=c("together","pmonly","mmonly","separate"),trim=0.02,baseline=-1,log.scalefactors=FALSE)

Arguments

X 	 A matrix. The columns of which are to be normalized.

abatch 	 An AffyBatch

type 	 A parameter controlling how normalization is applied to the Affybatch.

trim 	 How much to trim from the top and bottom before computing the mean when using the scaling normalization.

baseline 	 Index of array to use as baseline, negative values (-1,-2,-3,-4) control different baseline selection methods.

log.scalefactors 	 Compute the scale factors based on log2 transformed data.

Details

These function carries out scaling normalization of expression values.

Value

A normalized ExpressionSet.

Author(s)

Ben Bolstad, bmb@bmbolstad.com

See Also

normalize

Examples


data(affybatch.example)
normalize.AffyBatch.scaling(affybatch.example)

[Package affyPLM version 1.16.0 Index]