preprocess {affyPLM}	R Documentation

Background correct and Normalize

Description

This function pre-processes an AffyBatch.

Usage


preprocess(object,subset=NULL, normalize=TRUE,background=TRUE,background.method="RMA.2",normalize.method="quantile",background.param = list(),normalize.param=list(),verbosity.level=0)

Arguments

object 	 an AffyBatch

subset 	 a vector with the names of probesets to be used. If NULL then all probesets are used.

normalize 	 logical value. If TRUE normalize data using quantile normalization

background 	 logical value. If TRUE background correct using RMA background correction

background.method 	 name of background method to use.

normalize.method 	 name of normalization method to use.

background.param 	 list of parameters for background correction methods

normalize.param 	 list of parameters for normalization methods

verbosity.level 	 An integer specifying how much to print out. Higher values indicate more verbose. A value of 0 will print nothing

Details

This function carries out background correction and normalization pre-processing steps. It does not summarize to produce gene expression measures. All the same pre-processing methods supplied by threestep are supported by this function.

Value

An AffyBatch

Author(s)

Ben Bolstad bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley.

See Also

expresso, rma

Examples


data(affybatch.example)

# should be equivalent to the bg and norm of rma()
abatch.preprocessed <- preprocess(affybatch.example)


[Package affyPLM version 1.16.0 Index]