pseudo.coloring {affyPLM}	R Documentation

Coloring pseudo chip images

Description

These are routines used for coloring pseudo chip images.

Usage


  pseudoPalette(low = "white", high = c("green", "red"), mid = NULL,k =50) 
  pseudoColorBar(x, horizontal = TRUE, col = heat.colors(50), scale = 1:length(x),k = 11, log.ticks=FALSE,...)
  

Arguments

low 	 color at low end of scale

high 	 color at high end of scale

mid 	 color at exact middle of scale

k 	 number of colors to have

x 	 A data series

horizontal 	 If TRUE then color bar is to be draw horizontally

col 	 colors for color bar

scale 	 tickmarks for x if x is not numeric

log.ticks 	 use a log type transformation to assign the colors

... 	 additional parameters to plotting routine

Details

Adapted from similar tools in maPlots pacakge.

Author(s)

Ben Bolstad bmb@bmbolstad.com

See Also

AffyBatch, read.affybatch

[Package affyPLM version 1.16.0 Index]