rmaPLM {affyPLM}	R Documentation

Fit a RMA to Affymetrix Genechip Data as a PLMset

Description

This function converts an AffyBatch into an PLMset by fitting a multichip model. In particular we concentrate on the RMA model.

Usage


rmaPLM(object,subset=NULL,normalize=TRUE,background=TRUE,background.method="RMA.2",normalize.method="quantile",background.param = list(),normalize.param=list(),output.param=list(),model.param=list(),verbosity.level=0)

Arguments

object 	 an AffyBatch

subset 	 a vector with the names of probesets to be used. If NULL then all probesets are used.

normalize 	 logical value. If TRUE normalize data using quantile normalization

background 	 logical value. If TRUE background correct using RMA background correction

background.method 	 name of background method to use.

normalize.method 	 name of normalization method to use.

background.param 	 A list of parameters for background routines

normalize.param 	 A list of parameters for normalization routines

output.param 	 A list of parameters controlling optional output from the routine.

model.param 	 A list of parameters controlling model procedure

verbosity.level 	 An integer specifying how much to print out. Higher values indicate more verbose. A value of 0 will print nothing

Details

This function fits the RMA as a Probe Level Linear models to all the probesets in an AffyBatch.

Value

An PLMset

Author(s)

Ben Bolstad bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California,

Irizarry RA, Bolstad BM, Collin F, Cope LM, Hobbs B and Speed TP (2003) Summaries of Affymetrix GeneChip probe level data Nucleic Acids Research 31(4):e15

Bolstad, BM, Irizarry RA, Astrand, M, and Speed, TP (2003) A Comparison of Normalization Methods for High Density Oligonucleotide Array Data Based on Bias and Variance. Bioinformatics 19(2):185-193

See Also

expresso, rma, threestep,fitPLM, threestepPLM

Examples


# A larger example testing weight image function
data(Dilution)
## Not run: Pset <- rmaPLM(Dilution,output.param=list(weights=TRUE))
## Not run: image(Pset)

[Package affyPLM version 1.16.0 Index]