threestepPLM {affyPLM}	R Documentation

Three Step expression measures returned as a PLMset

Description

This function converts an AffyBatch into an PLMset using a three step expression measure.

Usage


threestepPLM(object,subset=NULL, normalize=TRUE,background=TRUE,background.method="RMA.2",normalize.method="quantile",summary.method="median.polish",background.param = list(),normalize.param=list(),output.param=list(), model.param=list(),verbosity.level=0)

Arguments

object 	 an AffyBatch

subset 	 a vector with the names of probesets to be used. If NULL then all probesets are used.

normalize 	 logical value. If TRUE normalize data using quantile normalization

background 	 logical value. If TRUE background correct using RMA background correction

background.method 	 name of background method to use.

normalize.method 	 name of normalization method to use.

summary.method 	 name of summary method to use.

background.param 	 list of parameters for background correction methods

normalize.param 	 list of parameters for normalization methods

output.param 	 list of parameters for output methods

model.param 	 list of parameters for model methods

verbosity.level 	 An integer specifying how much to print out. Higher values indicate more verbose. A value of 0 will print nothing

Details

This function computes the expression measure using threestep methods. It returns a PLMset. The most important difference is that the PLMset allows you to access the residuals which the threestep function does not do.

Value

An PLMset

Author(s)

Ben Bolstad bmb@bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley.

See Also

expresso, rma, threestep, rmaPLM, fitPLM

Examples


data(affybatch.example)

# should be equivalent to rma()
## Not run: eset <- threestepPLM(affybatch.example)

[Package affyPLM version 1.16.0 Index]